requirements.txt
fastapi
motor
pydantic
passlib
python-jose
python-dotenv
bcrypt
shortuuid
pymongo
google-auth
google-api-python-client